+++ flutter_easy_dialogs/lib/src/core/i_easy_dialog_manager_registry.dart
@@ -2,7 +2,7 @@ import 'package:flutter_easy_dialogs/src/core/easy_dialog_manager.dart';
 import 'package:flutter_easy_dialogs/src/easy_dialog_manager_controller/easy_dialog_manager_controller.dart';
 
 /// Class that is responsible for registering [EasyDialogManager].
-abstract class IEasyDialogsManagerRegistrar {
+abstract class IEasyDialogManagerRegistry {
   /// Register manager.
   void register<M extends EasyDialogManager>(
     EasyDialogManagerFactory<M> factory,